convert times to correct units

+++ chaos/include/Sequence.hpp
@@ -35,6 +35,8 @@ namespace Chaos {
    * The valid keys within each sequence step are the following:
    * - event: The type of signal event.
    * - command: The command that should be sent.
+   * 
+   * \todo: Use the chrono library to ensure our time units are consistent
    */
   class Sequence {
   protected:
@@ -55,8 +57,8 @@ namespace Chaos {
   public:
     Sequence(Controller& c);
 
-    static void setPressTime(unsigned int time);
-    static void setReleaseTime(unsigned int time);
+    static void setPressTime(double time);
+    static void setReleaseTime(double time);
 
     /**
      * \brief Directly add an individual event to the sequence stack
